About the calculator
This is a calculator I created in Excel  used in my previous job. I created this for fun and then let one of my coworkers use after explaining. It helped her with explaining final bill breakdowns for her. I want to create this same exact program in Python. I'm currently doing so now.